refactor: extract all API calls to a single file
